requirements.txt
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
tweepy==4.14.0
